@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_levensthein.
  *
- * (c) 2012-2016 The MetaModels team.
+ * (c) 2012-2018 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,7 +13,8 @@
  * @package    MetaModels
  * @subpackage Tests
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2012-2016 The MetaModels team.
+ * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
+ * @copyright  2012-2018 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_levensthein/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
@@ -25,7 +26,7 @@ use MetaModels\Attribute\Levensthein\AttributeLevensthein;
 /**
  * Unit tests to test class Decimal.
  */
-class LevenstheinAttributeTest extends \PHPUnit_Framework_TestCase
+class LevenstheinAttributeTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * Mock a MetaModel.
@@ -37,11 +38,11 @@ class LevenstheinAttributeTest extends \PHPUnit_Framework_TestCase
      */
     protected function mockMetaModel($language, $fallbackLanguage)
     {
-        $metaModel = $this->getMock(
-            'MetaModels\MetaModel',
-            array(),
-            array(array())
-        );
+        $metaModel = $this
+            ->getMockBuilder('MetaModels\MetaModel')
+            ->setMethods(array())
+            ->setConstructorArgs(array(array()))
+            ->getMock();
 
         $metaModel
             ->expects($this->any())
