@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_levensthein.
  *
- * (c) 2012-2016 The MetaModels team.
+ * (c) 2012-2018 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,7 +13,8 @@
  * @package    MetaModels
  * @subpackage AttributeLevensthein
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2012-2016 The MetaModels team.
+ * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
+ * @copyright  2012-2018 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_levensthein/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
@@ -120,7 +121,7 @@ class SearchStringParser
         preg_match_all('/"[^"]+"|[\+\-]?[^ ]+\*?/', $searchString, $chunks);
 
         if ($omitLastToken) {
-            $this->partial = array_pop($chunks[0]);
+            $this->partial = utf8_strtolower(array_pop($chunks[0]));
         }
         if (empty($chunks)) {
             return;
@@ -131,7 +132,7 @@ class SearchStringParser
         }
 
         if (!$omitLastToken) {
-            $this->partial = array_pop($chunks[0]);
+            $this->partial = utf8_strtolower(array_pop($chunks[0]));
         }
     }
 
